Espresso Start
Trying to successfully commit from desktop. Using multiple git accounts
should not be hell like this. My god, using stuff made by other people
is hell.

I wrote a skeleton of a parser. It's not much, but it *should* work. It
parses a json file into a vague "Action" class.