Update EspressoFileModifier.java
Changed an unnecessary change that was complicating the code. This class must implement FileModifierInterface not extend FileModifier. There is no need at all for the FileModifier class. I made the interface so that all you need is the interface to get it working correctly. Please see my UiAutomator code for an example as to how it all fits together.