Added the AbstractChecker
AbstractChecker is used to check that the input JSON file is correct.
Make sure to read all comments to make sure that you are using it
correctly. All you have to do to make a checker for your framework is
make a class that extends this class and override the check[action]
methods, the core of the checker is in this AbstractChecker so iteration
and returning are handled there.