Updated AppiumLive
Appium Live still works by using ids primarily, but will run by xpath if needed due to the generic-ness of "android:id" based elements. This allows us to specifically click on a mobile element based on any text the button has, its index, and related parameters. 
The xpath will always be created using the JSON parameters. If an xpath created cannot be used, this means there is something faulty with the JSON file, as the xpath will make use of parameters that should be explicity visible in either UIAutomator or Appium Inspector. 

1)In the future, it would definitely be worthwhile to specify bounds in the JSON file to help remove ambiguity between things such as icon buttons inside of other buttons which may have identical xpaths not including their bounds/location on the screen.
2)Also worth including would be original screen size for JSON file to make it so x and y coordinates can be mapped properly for things such as swipes.