Added Appium Checker
Appium Checker checks for everything used by the FileMode and LiveMode. After doing this, I'm pretty sure our group is using the text field incorrectly. We should probably use a different parameter than "text" to refer to the text we wish to type into a textfield on screen. The text field as its mentioned by UIAutomatorViewer and the Appium Inspector seems to refer to actual text on the buttons so this is more for locating buttons. We shouldnt mix and match like that.