Update AppiumFunctions.java
typing functions not added because a template is already more or less available in AppiumLive, and committing dead redundant template code is pointless. Typing will not be available for this sprint due to issues with x and y coordinates in JSON files. May require original screen size for proper scaling on next sprint.