Kinda useless parsing back to JSON files
I sorta technically did parse code back to a JSON file. But I did it in
a dead language instead of Java. I'm pretty sure it shaved a couple
hours off of my time, I swear it's not a spite thing. I mostly wanted
to see if I could do it in snobol, since this whole project is pretty
much made to play to snobol's strengths as a language.